Extract string utilities into separate file
