+++ src/construction.cpp
@@ -11,7 +11,6 @@
 #include "action.h"
 #include "avatar.h"
 #include "calendar.h"
-#include "cata_utility.h"
 #include "colony.h"
 #include "color.h"
 #include "construction_category.h"
@@ -48,6 +47,7 @@
 #include "string_formatter.h"
 #include "string_id.h"
 #include "string_input_popup.h"
+#include "string_utils.h"
 #include "translations.h"
 #include "trap.h"
 #include "ui_manager.h"
