+++ src/debug_menu.cpp
@@ -81,6 +81,7 @@
 #include "string_formatter.h"
 #include "string_id.h"
 #include "string_input_popup.h"
+#include "string_utils.h"
 #include "trait_group.h"
 #include "translations.h"
 #include "type_id.h"
