+++ src/mapgen.cpp
@@ -52,7 +52,6 @@
 #include "omdata.h"
 #include "optional.h"
 #include "options.h"
-#include "output.h"
 #include "overmap.h"
 #include "overmapbuffer.h"
 #include "player.h"
@@ -60,6 +59,7 @@
 #include "rng.h"
 #include "string_formatter.h"
 #include "string_id.h"
+#include "string_utils.h"
 #include "submap.h"
 #include "text_snippets.h"
 #include "tileray.h"
