+++ src/martialarts.cpp
@@ -32,6 +32,7 @@
 #include "skill.h"
 #include "string_formatter.h"
 #include "string_id.h"
+#include "string_utils.h"
 #include "translations.h"
 #include "ui_manager.h"
 #include "value_ptr.h"
