+++ src/messages.cpp
@@ -15,6 +15,7 @@
 #include "point.h"
 #include "string_formatter.h"
 #include "string_input_popup.h"
+#include "string_utils.h"
 #include "translations.h"
 #include "ui_manager.h"
