+++ src/options.cpp
@@ -31,6 +31,7 @@
 #include "sounds.h"
 #include "string_formatter.h"
 #include "string_input_popup.h"
+#include "string_utils.h"
 #include "translations.h"
 #include "ui_manager.h"
 #include "worldfactory.h"
