+++ src/requirements.cpp
@@ -28,6 +28,7 @@
 #include "point.h"
 #include "string_formatter.h"
 #include "string_id.h"
+#include "string_utils.h"
 #include "translations.h"
 #include "visitable.h"
