+++ src/safemode_ui.cpp
@@ -26,6 +26,7 @@
 #include "point.h"
 #include "string_formatter.h"
 #include "string_input_popup.h"
+#include "string_utils.h"
 #include "translations.h"
 #include "ui_manager.h"
