+++ src/veh_interact.cpp
@@ -51,6 +51,7 @@
 #include "string_formatter.h"
 #include "string_id.h"
 #include "string_input_popup.h"
+#include "string_utils.h"
 #include "tileray.h"
 #include "translations.h"
 #include "ui.h"
